Update list_group_utils.ml
Added function mappings to allow use of labels. Sort module is deprecated as it was just a wrapping around Sortx functions to take a single sorting function. Added type to make it easier to specify that a named key will be the same as the key without having to duplicate the function.